fix: cleaned up some classes in crypto
