+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Crypto/CryptoOperation.cs
@@ -1,0 +1,9 @@
+namespace Plugin.Maui.Biometric;
+
+public enum CryptoOperation
+{
+    Sign,
+    Verify,
+    Encrypt,
+    Decrypt
+}
