+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Crypto/IBiometricCrypto.cs
@@ -45,42 +45,3 @@ public interface IBiometricCrypto
     /// </summary>
     Task<bool> KeyExistsAsync(string keyId);
 }
-
-
-public sealed class SecureCryptoResponse
-{
-    public bool WasSuccessful { get; set; }
-    public byte[]? OutputData { get; set; }
-    public string? ErrorMessage { get; set; }
-}
-
-public class CryptoKeyOptions
-{
-    /// <summary>
-    /// Operation type (Encrypt, Decrypt, Sign, Verify, MAC)
-    /// </summary>
-    public CryptoOperation Operation { get; set; }
-
-    /// <summary>
-    /// Optional algorithm (e.g., AES, RSA, ECDSA)
-    /// </summary>
-    public string? Algorithm { get; set; }
-
-    /// <summary>
-    /// Optional key size in bits
-    /// </summary>
-    public int? KeySize { get; set; }
-
-    /// <summary>
-    /// Optional flag for requiring biometric authentication to use the key
-    /// </summary>
-    public bool RequireUserAuthentication { get; set; } = true;
-}
-
-public enum CryptoOperation
-{
-    Sign,
-    Verify,
-    Encrypt,
-    Decrypt
-}
